py_ant_maze/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-ant-maze"
version = "0.1.1"
description = "Minimal utilities for maze definitions"
readme = "README.md"
authors = [
    {name = "Yihao Liu", email = "yihao.jhu@gmail.com"}
]
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["maze", "navigation", "grid", "environment"]
dependencies = [
    "PyYAML>=6.0",
    "numpy>=1.21",
    "Pillow>=9.0",
]
requires-python = ">=3.7"

[project.urls]
"Homepage" = "https://github.com/brain-sim/py-ant-maze"
"Bug Tracker" = "https://github.com/brain-sim/py-ant-maze/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["py_ant_maze*"]
exclude = ["tests*", "examples*"]
